backend/requirements.txt
# Django core
django==5.1.3
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2

# Database
psycopg[binary]==3.2.3

# Task queue
celery==5.4.0
redis==5.2.0
django-redis==5.4.0

# Storage
django-storages[boto3]==1.14.4
boto3==1.35.63

# Image/PDF processing
Pillow==11.0.0
pdf2image==1.17.0
weasyprint==63.1

# API/Security
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework-simplejwt==5.3.1

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
factory_boy==3.3.1
freezegun==1.5.1

# Type checking
mypy==1.13.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1

# Code formatting/linting
black==24.10.0
isort==5.13.2
ruff==0.8.1
pre-commit==4.0.1
djlint==1.36.1